712 remove renderinglibrary reference from the gum project (#715)
* More refactoring, removing RenderingLibrary from Gum, and it actually compiles!

* Plugins run!

* Fixed selection of instances preferring the current item first before checking others.

* Removed duplicate interfaces causing clip children to fail.

* Finished removing RenderingLibrary from Gum
