+++ Gum/Graphics/Animation/Content/AnimationChainListSave.cs
@@ -1,10 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Xml.Serialization;
-using AnimationChainList = Gum.Graphics.Animation.AnimationChainList;
 using ToolsUtilities;
-using Gum.Graphics.Animation;
-using TimeMeasurementUnit = Gum.Graphics.Animation.TimeMeasurementUnit;
 
 namespace Gum.Content.AnimationChain
 {
@@ -176,184 +173,11 @@ namespace Gum.Content.AnimationChain
         //}
 
 
-        public AnimationChainList ToAnimationChainList(string contentManagerName)
-        {
-
-            return ToAnimationChainList(contentManagerName, true);
-        }
-
-
-        public AnimationChainList ToAnimationChainList(string contentManagerName, bool throwError)
-        {
-            mToRuntimeErrors.Clear();
-
-            AnimationChainList list = new AnimationChainList();
-
-            list.FileRelativeTextures = FileRelativeTextures;
-            list.TimeMeasurementUnit = TimeMeasurementUnit;
-            list.Name = mFileName;
-
-            string oldRelativeDirectory = FileManager.RelativeDirectory;
-
-            try
-            {
-                if (this.FileRelativeTextures)
-                {
-                    FileManager.RelativeDirectory = FileManager.GetDirectory(mFileName);
-                }
-
-                foreach (AnimationChainSave animationChain in this.AnimationChains)
-                {
-                    try
-                    {
-                        Gum.Graphics.Animation.AnimationChain newChain = null;
-
-                        newChain = animationChain.ToAnimationChain(contentManagerName, this.TimeMeasurementUnit, this.CoordinateType);
-
-                        newChain.IndexInLoadedAchx = list.Count;
-
-                        newChain.ParentAchxFileName = mFileName;
-
-                        list.Add(newChain);
-
-                    }
-                    catch (Exception e)
-                    {
-                        mToRuntimeErrors.Add(e.ToString());
-                        if (throwError)
-                        {
-                            throw new Exception("Error loading AnimationChain", e);
-                        }
-                    }
-                }
-            }
-            finally
-            {
-                FileManager.RelativeDirectory = oldRelativeDirectory;
-            }
-
-            return list;
-        }
 
         public void Dispose()
         {
             // do nothing, just need this to add it to the loader manager
         }
-
-
-        //AnimationChainList ToAnimationChainList(string contentManagerName, TextureAtlas textureAtlas, bool throwError)
-        //{
-        //    mToRuntimeErrors.Clear();
-
-        //    AnimationChainList list = new AnimationChainList();
-
-        //    list.FileRelativeTextures = FileRelativeTextures;
-        //    list.TimeMeasurementUnit = TimeMeasurementUnit;
-        //    list.Name = mFileName;
-
-        //    string oldRelativeDirectory = FileManager.RelativeDirectory;
-
-        //    try
-        //    {
-        //        if (this.FileRelativeTextures)
-        //        {
-        //            FileManager.RelativeDirectory = FileManager.GetDirectory(mFileName);
-        //        }
-
-        //        foreach (AnimationChainSave animationChain in this.AnimationChains)
-        //        {
-        //            try
-        //            {
-        //                FlatRedBall.Graphics.Animation.AnimationChain newChain = null;
-
-        //                if (textureAtlas == null)
-        //                {
-        //                    newChain = animationChain.ToAnimationChain(contentManagerName, this.TimeMeasurementUnit, this.CoordinateType);
-        //                }
-        //                else
-        //                {
-        //                    newChain = animationChain.ToAnimationChain(textureAtlas, this.TimeMeasurementUnit);
-        //                }
-        //                newChain.mIndexInLoadedAchx = list.Count;
-
-        //                newChain.ParentAchxFileName = mFileName;
-
-        //                list.Add(newChain);
-
-        //            }
-        //            catch (Exception e)
-        //            {
-        //                mToRuntimeErrors.Add(e.ToString());
-        //                if (throwError)
-        //                {
-        //                    throw new Exception("Error loading AnimationChain", e);
-        //                }
-        //            }
-        //        }
-        //    }
-        //    finally
-        //    {
-        //        FileManager.RelativeDirectory = oldRelativeDirectory;
-        //    }
-
-        //    return list;
-
-
-        //}
-
-
-        //public AnimationChainList ToAnimationChainList(Graphics.Texture.TextureAtlas textureAtlas)
-        //{
-        //    return ToAnimationChainList(null, textureAtlas, true);
-        //}
-
-
-        //private void MakeRelative(string fileName)
-        //{
-        //    string oldRelativeDirectory = FileManager.RelativeDirectory;
-
-        //    string newRelativeDirectory = FileManager.GetDirectory(fileName);
-        //    FileManager.RelativeDirectory = newRelativeDirectory;
-
-        //    foreach (AnimationChainSave acs in AnimationChains)
-        //    {
-        //        acs.MakeRelative();
-
-        //    }
-
-        //    FileManager.RelativeDirectory = oldRelativeDirectory;
-        //}
-
-
-        //private static AnimationChainListSave DeserializeManually(string fileName)
-        //{
-        //    AnimationChainListSave toReturn = new AnimationChainListSave();
-        //    System.Xml.Linq.XDocument xDocument = null;
-
-        //    using (var stream = FileManager.GetStreamForFile(fileName))
-        //    {
-        //        xDocument = System.Xml.Linq.XDocument.Load(stream);
-        //    }
-
-        //    System.Xml.Linq.XElement foundElement = null;
-
-        //    foreach (var element in xDocument.Elements())
-        //    {
-        //        if (element.Name.LocalName == "AnimationChainArraySave")
-        //        {
-        //            foundElement = element;
-        //            break;
-        //        }
-        //    }
-
-        //    LoadFromElement(toReturn, foundElement);
-
-        //    return toReturn;
-        //}
-
-
-
-
 
         #endregion
     }
