+++ Gum/Gum.csproj
@@ -75,12 +75,6 @@
 		</Reference>
 	</ItemGroup>
 	<ItemGroup>
-		<Compile Include="..\GumRuntime\BbCodeParser.cs">
-			<Link>Wireframe\BbCodeParser.cs</Link>
-		</Compile>
-		<Compile Include="..\MonoGameGum\Localization\ILocalizationService.cs">
-			<Link>Managers\ILocalizationService.cs</Link>
-		</Compile>
 		<None Update="Content\Fonts\Font18Arial_o1.fnt">
 			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 		</None>
@@ -94,10 +88,8 @@
 		<ProjectReference Include="..\GumCommon\GumCommon.csproj" />
 		<ProjectReference Include="..\GumDataTypes\GumDataTypesNet6.csproj" />
 		<ProjectReference Include="..\InputLibrary\InputLibrary.csproj" />
-		<ProjectReference Include="..\RenderingLibrary\RenderingLibrary.csproj" />
 		<ProjectReference Include="..\ToolsUtilities\ToolsUtilitiesStandard.csproj" />
 		<ProjectReference Include="..\WpfDataUi\WpfDataUi.csproj" />
-		<ProjectReference Include="..\XnaAndWinforms\XnaAndWinforms.csproj" />
 	</ItemGroup>
 	<ItemGroup>
 		<Resource Include="Plugins\InternalPlugins\TreeView\Content\behavior.png" />
@@ -317,6 +309,7 @@
 	</ItemGroup>
 	<ItemGroup>
 		<Folder Include="Converters\" />
+		<Folder Include="Graphics\Animation\Content\" />
 		<Folder Include="Libraries\Xna\" />
 		<Folder Include="Plugins\InternalPlugins\ScreenshotPlugin\" />
 	</ItemGroup>
@@ -374,6 +367,9 @@
 		<Compile Remove="DataTypes\RecursiveVariableFinder.cs" />
 		<Compile Remove="DataTypes\StateSaveExtensionMethods.cs" />
 		<Compile Remove="DataTypes\VariableSaveExtensionMethods.cs" />
+		<Compile Remove="Graphics\Animation\Content\AnimationChainListSave.cs" />
+		<Compile Remove="Graphics\Animation\Content\AnimationChainSave.cs" />
+		<Compile Remove="Graphics\Animation\Content\AnimationFrameSave.cs" />
 		<Compile Remove="Managers\ObjectFinder.cs" />
 		<Compile Remove="Managers\StandardElementsManager.cs" />
 		<Compile Remove="Mvvm\ViewModel.cs" />
@@ -436,7 +432,10 @@
 		<Compile Remove="objNetFramework\x86\Debug\Plugins\InternalPlugins\VariableGrid\VariableRemoveButton.g.i.cs" />
 		<Compile Remove="Plugins\InternalPlugins\AlignmentButtons\AlignmentControl.xaml.cs" />
 		<Compile Remove="RenderingLibrary\Blend.cs" />
+		<Compile Remove="RenderingLibrary\BlendExtensions.cs" />
+		<Compile Remove="Wireframe\CustomSetPropertyOnRenderable.cs" />
 		<Compile Remove="Wireframe\ElementWithState.cs" />
+		<Compile Remove="Wireframe\RuntimeObjectCreator.cs" />
 		<Page Remove="Plugins\InternalPlugins\AlignmentButtons\AlignmentControl.xaml" />
 	</ItemGroup>
 </Project>
