+++ Gum/Wireframe/WireframeObjectManager.RepresentationCreation.cs
@@ -113,245 +113,6 @@ public partial class WireframeObjectManager
         }
     }
 
-    private GraphicalUiElement CreateRepresentationForInstance(InstanceSave instance, InstanceSave parentInstance, List<ElementWithState> elementStack, GraphicalUiElement container)
-    {
-        IPositionedSizedObject newIpso = null;
-
-        GraphicalUiElement graphicalElement = newIpso as GraphicalUiElement;
-        var baseElement = ObjectFinder.Self.GetElementSave(instance.BaseType);
-
-        string type = instance.BaseType;
-
-        var renderable = PluginManager.Self.CreateRenderableForType(type);
-
-        if(renderable != null)
-        {
-            graphicalElement = new GraphicalUiElement(null, container);
-            ElementSave instanceBase = ObjectFinder.Self.GetElementSave(instance.BaseType);
-            graphicalElement.SetContainedObject(renderable);
-            graphicalElement.Tag = instance;
-            graphicalElement.Name = instance.Name;
-            graphicalElement.Component.Tag = instance;
-        }
-
-
-        else if (type == "Sprite" || type == "ColoredRectangle" || type == "NineSlice" || type == "Text" || type == "Circle" || type == "Rectangle")
-        {
-            graphicalElement = new GraphicalUiElement(null, container);
-            ElementSave instanceBase = ObjectFinder.Self.GetElementSave(type);
-            graphicalElement.CreateGraphicalComponent(instanceBase, null);
-            graphicalElement.Tag = instance;
-            graphicalElement.Name = instance.Name;
-            graphicalElement.Component.Tag = instance;
-
-            if (type == "Text")
-            {
-                (graphicalElement.RenderableComponent as Text).RenderBoundary = ProjectManager.Self.GeneralSettingsFile.ShowTextOutlines;
-                if(SelectedState.Self.SelectedStateSave != null)
-                {
-                    var instanceElement = ObjectFinder.Self.GetElementSave(instance);
-                    if (instanceElement != null)
-                    {
-                        elementStack.Add(new ElementWithState(instanceElement));
-                    }
-                    var rfv = new RecursiveVariableFinder(elementStack);
-
-
-                    var forcedValues = new StateSave();
-
-                    void TryAddForced(string variableName)
-                    {
-                        var value = rfv.GetValueByBottomName(variableName);
-                        if (value != null)
-                        {
-                            forcedValues.SetValue(variableName, value);
-                        }
-                    }
-
-                    TryAddForced("Font");
-                    TryAddForced("FontSize");
-                    TryAddForced("OutlineThickness");
-                    TryAddForced("UseFontSmoothing");
-                    TryAddForced("IsItalic");
-                    TryAddForced("IsBold");
-
-                    StateSave stateSave = SelectedState.Self.SelectedStateSave;
-
-                    // If the user has a category selected but no state in the category, then use the default:
-                    if (stateSave == null && SelectedState.Self.SelectedStateCategorySave != null)
-                    {
-                        stateSave = SelectedState.Self.SelectedElement.DefaultState;
-                    }
-
-
-
-                    _fontManager.ReactToFontValueSet(instance, GumState.Self.ProjectState.GumProjectSave, stateSave, forcedValues);
-                }
-            }
-
-        }
-        else if (instance.IsComponent())
-        {
-            newIpso = CreateRepresentationsForInstanceFromComponent(instance, elementStack, parentInstance, container, ObjectFinder.Self.GetComponent(instance.BaseType));
-        }
-        else
-        {
-            // Make sure the base type is valid.
-            // This could be null if a base type changed
-            // its name but the derived wasn't updated, or 
-            // if someone screwed with the XML files.  Who knows...
-
-            if (baseElement != null)
-            {
-                graphicalElement = new GraphicalUiElement(null, container);
-
-                CreateRectangleFor(instance, elementStack, graphicalElement);
-            }
-        }
-
-
-
-        if (newIpso != null && (newIpso is GraphicalUiElement) == false)
-        {
-            graphicalElement = new GraphicalUiElement(newIpso as IRenderable, container);
-        }
-        else if (newIpso is GraphicalUiElement)
-        {
-            graphicalElement = newIpso as GraphicalUiElement;
-        }
-
-        if(graphicalElement != null)
-        {
-            if (baseElement != null)
-            {
-                graphicalElement.ElementSave = baseElement;
-            }
-
-            graphicalElement.AddExposedVariablesRecursively(baseElement);
-
-            var selectedState = SelectedState.Self.SelectedStateSave;
-            if (selectedState == null)
-            {
-                selectedState = SelectedState.Self.SelectedElement.DefaultState;
-            }
-
-            RecursiveVariableFinder rvf = new DataTypes.RecursiveVariableFinder(selectedState);
-
-            string guide = rvf.GetValue<string>("Guide");
-            string parent = rvf.GetValue<string>(instance.Name + ".Parent");
-
-            SetGuideParent(container, graphicalElement, guide);
-
-            if (baseElement != null)
-            {
-                graphicalElement.AddStatesAndCategoriesRecursivelyToGue(baseElement);
-
-                // for performance reasons, we'll suspend the layout here:
-                graphicalElement.SuspendLayout();
-                {
-                    graphicalElement.SetVariablesRecursively(baseElement, baseElement.DefaultState);
-                }
-                graphicalElement.ResumeLayout();
-            }
-        }
-
-        return graphicalElement;
-    }
-
-    private GraphicalUiElement CreateRepresentationsForInstanceFromComponent(InstanceSave instance, 
-        List<ElementWithState> elementStack, InstanceSave parentInstance, GraphicalUiElement parentIpso, 
-        ComponentSave baseComponentSave)
-    {
-        StandardElementSave ses = ObjectFinder.Self.GetRootStandardElementSave(instance);
-
-        GraphicalUiElement rootIpso = null;
-        if (ses != null)
-        {
-            rootIpso = new GraphicalUiElement(null, parentIpso);
-
-            string type = ses.Name;
-
-            var renderable = PluginManager.Self.CreateRenderableForType(type);
-
-            if(renderable != null)
-            {
-                rootIpso.SetContainedObject(renderable);
-                rootIpso.Tag = instance;
-                rootIpso.Name = instance.Name;
-                rootIpso.Component.Tag = instance;
-            }
-
-            else if (type == "Sprite" || type == "ColoredRectangle" || type == "NineSlice" || type == "Text" || type == "Circle" || type == "Rectangle")
-            {
-                ElementSave instanceBase = ObjectFinder.Self.GetElementSave(instance.BaseType);
-                rootIpso.CreateGraphicalComponent(instanceBase, null);
-                rootIpso.Tag = instance;
-                rootIpso.Name = instance.Name;
-                rootIpso.Component.Tag = instance;
-
-                if(type == "Text")
-                {
-                    (rootIpso.RenderableComponent as Text).RenderBoundary = ProjectManager.Self.GeneralSettingsFile.ShowTextOutlines;
-                }
-
-            }
-            else
-            {
-                // give the plugin manager a shot at it:
-
-                CreateRectangleFor(instance, elementStack, rootIpso);
-            }
-
-            var selectedState = SelectedState.Self.SelectedStateSave;
-
-            if (selectedState == null)
-            {
-                selectedState = SelectedState.Self.SelectedElement.DefaultState;
-            }
-
-            RecursiveVariableFinder rvf = new DataTypes.RecursiveVariableFinder(selectedState);
-
-            string guide = rvf.GetValue<string>("Guide");
-            SetGuideParent(parentIpso, rootIpso, guide);
-
-            ElementWithState elementWithStateForNewInstance = new ElementWithState(baseComponentSave);
-            var tempRvf = new DataTypes.RecursiveVariableFinder(instance, elementStack);
-            var state = tempRvf.GetValue("State") as string;
-            elementWithStateForNewInstance.StateName = state;
-
-            foreach (var category in baseComponentSave.Categories)
-            {
-                elementWithStateForNewInstance.CategorizedStates.Add(category.Name, tempRvf.GetValue<string>(category.Name + "State"));
-            }
-
-            // This seems wrong - we shouldn't be setting the instance name on the elementState for the instance. It should be one above:
-            //elementWithStateForNewInstance.InstanceName = instance.Name;
-            elementStack.Last().InstanceName = instance.Name;
-
-            // Does this element already exist?
-            bool alreadyExists = elementStack.Any(item => item.Element == elementWithStateForNewInstance.Element);
-            if(!alreadyExists)
-            {
-                elementStack.Add(elementWithStateForNewInstance);
-
-                foreach (InstanceSave internalInstance in baseComponentSave.Instances)
-                {
-                    // let's make sure we don't recursively create the same instance causing a stack overflow:
-                    elementWithStateForNewInstance.InstanceName = internalInstance.Name;
-
-                    GraphicalUiElement createdIpso = CreateRepresentationForInstance(internalInstance, instance, elementStack, rootIpso);
-
-                }
-            }
-
-            SetUpParentRelationship(rootIpso.ContainedElements, elementStack);
-
-            elementStack.Remove(elementStack.FirstOrDefault(item => item.Element == baseComponentSave));
-        }
-
-        return rootIpso;
-    }
-
     private void SetUpParentRelationship(IEnumerable<GraphicalUiElement> elements, List<ElementWithState> elementStack)
     {
         // Now that we have created all instances, we can establish parent relationships
