+++ GumCommon/GumCommon.csproj
@@ -25,10 +25,14 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+		<Compile Include="..\GumRuntime\BbCodeParser.cs" Link="Runtime\BbCodeParser.cs" />
 		<Compile Include="..\GumRuntime\BindableGue.cs" Link="BindableGue.cs" />
 		<Compile Include="..\GumRuntime\ElementSaveExtensions.GumRuntime.cs" Link="Wireframe\ElementSaveExtensions.GumRuntime.cs" />
 		<Compile Include="..\GumRuntime\InstanceSaveExtensionMethods.GumRuntime.cs" Link="Wireframe\InstanceSaveExtensionMethods.GumRuntime.cs" />
 		<Compile Include="..\GumRuntime\InteractiveGue.cs" Link="InteractiveGue.cs" />
+		<Compile Include="..\Gum\Graphics\Animation\Content\AnimationChainListSave.cs" Link="Animation\AnimationChainListSave.cs" />
+		<Compile Include="..\Gum\Graphics\Animation\Content\AnimationChainSave.cs" Link="Animation\AnimationChainSave.cs" />
+		<Compile Include="..\Gum\Graphics\Animation\Content\AnimationFrameSave.cs" Link="Animation\AnimationFrameSave.cs" />
 		<Compile Include="..\Gum\Managers\ObjectFinder.cs" Link="Managers\ObjectFinder.cs" />
 		<Compile Include="..\Gum\Mvvm\ViewModel.cs" Link="Mvvm\ViewModel.cs" />
 		<Compile Include="..\RenderingLibrary\Content\IContentLoader.cs" Link="Content\IContentLoader.cs" />
@@ -50,6 +54,7 @@
 		<Folder Include="Content\" />
 		<Folder Include="Managers\" />
 		<Folder Include="Mvvm\" />
+		<Folder Include="Animation\" />
 		<Folder Include="RenderingLibrary\" />
 	</ItemGroup>
