+++ GumCommon/Localization/ILocalizationService.cs
