+++ GumRuntime/BbCodeParser.cs
@@ -1,9 +1,6 @@
 
 
-using System.Collections.Generic;
 using System.Text;
-using System;
-using System.Linq;
 
 namespace Gum.Wireframe
 {
