+++ Tool/EditorTabPlugin_XNA/Services/HighlightManager.cs
