+++ Tool/EditorTabPlugin_XNA/Services/ScrollBarControlLogic.cs
