+++ Tool/EditorTabPlugin_XNA/Services/ToolFontService.cs
@@ -23,7 +23,10 @@ namespace Gum.Managers
         }
         public void Initialize()
         {
-            var executingPath = Assembly.GetExecutingAssembly().Location;
+            // This is the plugin, which doesn't have the font...
+            //var executingPath = Assembly.GetExecutingAssembly().Location;
+            // Instead use Gum which is the entry assembly
+            var executingPath = Assembly.GetEntryAssembly().Location;
             var directory = Path.GetDirectoryName(executingPath);
 
             var fntFilePath = Path.Combine(directory, "Content/Fonts/Font18Arial_o1.fnt");
