+++ Tool/EditorTabPlugin_XNA/Views/OriginDisplay.cs
